refactor: update function signatures to use size_t and uint32_t, remove memcpy's from thash function

+++ fw/sphincs_hw/fips202.c
@@ -17,7 +17,7 @@ typedef struct {
     uint32_t squeeze_rem;
 } hw_keccak_ctx;
 
-static hw_keccak_ctx* get_ctx(uint64_t *s_inc) {
+static hw_keccak_ctx* get_ctx(uint32_t *s_inc) {
     return (hw_keccak_ctx*)s_inc;
 }
 
@@ -103,7 +103,7 @@ static void hw_squeeze(hw_keccak_ctx *ctx, uint8_t *output, size_t outlen) {
 }
 
 // SHAKE128
-void shake128_inc_init(uint64_t *s_inc) {
+void shake128_inc_init(uint32_t *s_inc) {
     hw_keccak_ctx *ctx = get_ctx(s_inc);
     ctx->buf_len = 0;
     ctx->rate_words = SHAKE128_RATE >> 3;
@@ -113,24 +113,24 @@ void shake128_inc_init(uint64_t *s_inc) {
     keccak_init(ctx->rate_words, KECCAK_MASK_DIS);
 }
 
-void shake128_inc_absorb(uint64_t *s_inc, const uint8_t *input, size_t inlen) {
+void shake128_inc_absorb(uint32_t *s_inc, const uint8_t *input, size_t inlen) {
     hw_absorb(get_ctx(s_inc), input, inlen);
 }
 
-void shake128_inc_finalize(uint64_t *s_inc) {
+void shake128_inc_finalize(uint32_t *s_inc) {
     hw_finalize(get_ctx(s_inc));
 }
 
-void shake128_inc_squeeze(uint8_t *output, size_t outlen, uint64_t *s_inc) {
+void shake128_inc_squeeze(uint8_t *output, size_t outlen, uint32_t *s_inc) {
     hw_squeeze(get_ctx(s_inc), output, outlen);
 }
 
-void shake128_absorb(uint64_t *s, const uint8_t *input, size_t inlen) {
+void shake128_absorb(uint32_t *s, const uint8_t *input, size_t inlen) {
     shake128_inc_init(s);
     shake128_inc_absorb(s, input, inlen);
 }
 
-void shake128_squeezeblocks(uint8_t *output, size_t nblocks, uint64_t *s) {
+void shake128_squeezeblocks(uint8_t *output, size_t nblocks, uint32_t *s) {
     hw_finalize(get_ctx(s));
     hw_squeeze(get_ctx(s), output, nblocks * SHAKE128_RATE);
 }
@@ -146,7 +146,7 @@ void shake128(uint8_t *output, size_t outlen, const uint8_t *input, size_t inlen
 }
 
 // SHAKE256
-void shake256_inc_init(uint64_t *s_inc) {
+void shake256_inc_init(uint32_t *s_inc) {
     hw_keccak_ctx *ctx = get_ctx(s_inc);
     ctx->buf_len = 0;
     ctx->rate_words = SHAKE256_RATE >> 3;
@@ -156,24 +156,24 @@ void shake256_inc_init(uint64_t *s_inc) {
     keccak_init(ctx->rate_words, KECCAK_MASK_DIS);
 }
 
-void shake256_inc_absorb(uint64_t *s_inc, const uint8_t *input, size_t inlen) {
+void shake256_inc_absorb(uint32_t *s_inc, const uint8_t *input, size_t inlen) {
     hw_absorb(get_ctx(s_inc), input, inlen);
 }
 
-void shake256_inc_finalize(uint64_t *s_inc) {
+void shake256_inc_finalize(uint32_t *s_inc) {
     hw_finalize(get_ctx(s_inc));
 }
 
-void shake256_inc_squeeze(uint8_t *output, size_t outlen, uint64_t *s_inc) {
+void shake256_inc_squeeze(uint8_t *output, size_t outlen, uint32_t *s_inc) {
     hw_squeeze(get_ctx(s_inc), output, outlen);
 }
 
-void shake256_absorb(uint64_t *s, const uint8_t *input, size_t inlen) {
+void shake256_absorb(uint32_t *s, const uint8_t *input, size_t inlen) {
     shake256_inc_init(s);
     shake256_inc_absorb(s, input, inlen);
 }
 
-void shake256_squeezeblocks(uint8_t *output, size_t nblocks, uint64_t *s) {
+void shake256_squeezeblocks(uint8_t *output, size_t nblocks, uint32_t *s) {
     hw_finalize(get_ctx(s));
     hw_squeeze(get_ctx(s), output, nblocks * SHAKE256_RATE);
 }
