+++ fw/sphincs_hw/hash.h
@@ -15,13 +15,13 @@ void prf_addr(unsigned char *out, const spx_ctx *ctx,
 #define gen_message_random SPX_NAMESPACE(gen_message_random)
 void gen_message_random(unsigned char *R, const unsigned char *sk_prf,
                         const unsigned char *optrand,
-                        const unsigned char *m, unsigned long long mlen,
+                        const unsigned char *m, size_t mlen,
                         const spx_ctx *ctx);
 
 #define hash_message SPX_NAMESPACE(hash_message)
-void hash_message(unsigned char *digest, uint64_t *tree, uint32_t *leaf_idx,
+void hash_message(unsigned char *digest, uint32_t *tree, uint32_t *leaf_idx,
                   const unsigned char *R, const unsigned char *pk,
-                  const unsigned char *m, unsigned long long mlen,
+                  const unsigned char *m, size_t mlen,
                   const spx_ctx *ctx);
 
 #endif
